refactor: move cipher validation to a separate module
